feat(cli): establish npm release boundary
Separate the production bin from repository CLI logic and close the package's accidental JavaScript import surface while preserving the current public commands.

Fail Maka Eval before trial startup when pinned Python, Docker, or machine prerequisites are unavailable, using the same credential-filtered preparation environment as the trial supervisor.

Generated-by: OpenAI Codex
