+++ packages/cli/package.json
@@ -4,8 +4,7 @@
   "license": "Apache-2.0",
   "private": true,
   "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
+  "exports": {},
   "bin": {
     "maka": "./dist/cli.js",
     "maka-agent": "./dist/cli.js"
