+++ packages/eval/src/harness-executor.ts
@@ -4,10 +4,11 @@ import { once } from 'node:events';
 import { createReadStream } from 'node:fs';
 import { chmod, lstat, mkdir, readFile, readdir, unlink, writeFile } from 'node:fs/promises';
 import { createServer, type Server, type Socket } from 'node:net';
-import { basename, delimiter, dirname, join, relative, resolve, sep } from 'node:path';
+import { basename, dirname, join, relative, resolve, sep } from 'node:path';
 import { createInterface } from 'node:readline';
 import { fileURLToPath } from 'node:url';
 import { decodeJsonObject, type ExperimentCell, type JsonObject } from './experiment.js';
+import { createHarnessPreparationEnvironment } from './harness-environment.js';
 import {
   MAKA_RUNTIME_ARTIFACT_PATH,
   MAKA_SUBJECT_STDERR_PATH,
@@ -358,7 +359,7 @@ async function startTrial(
     ...UNATTENDED_EXECUTION_ENVIRONMENT,
     ...egressExecutionEnvironment(options.egressProxy),
   };
-  const environment = preparationEnvironment(
+  const environment = createHarnessPreparationEnvironment(
     relayPath,
     [...subjectCredentialNames, ...cell.subject.credentials],
     options.preparationEnvironment,
@@ -587,48 +588,6 @@ function preparationCode(
   if (stage === 'spawn') return 'spawn-failed';
   if (stage === 'ready-decode') return 'invalid-ready';
   return 'exit-before-ready';
-}
-
-function preparationEnvironment(
-  relayPath: string,
-  subjectCredentialNames: readonly string[],
-  declared: readonly string[],
-  egressAllowedHost?: string,
-  networkPolicyPath?: string,
-): NodeJS.ProcessEnv {
-  const allowed = new Set([
-    'HOME',
-    'PATH',
-    'TMPDIR',
-    'TMP',
-    'TEMP',
-    'LANG',
-    'LC_ALL',
-    'SSL_CERT_FILE',
-    'SSL_CERT_DIR',
-    'REQUESTS_CA_BUNDLE',
-    'CURL_CA_BUNDLE',
-    'XDG_CACHE_HOME',
-    ...declared,
-  ]);
-  const credentials = new Set(subjectCredentialNames);
-  const inherited = Object.fromEntries(
-    [...allowed].flatMap((name) => {
-      const value = process.env[name];
-      return value === undefined || credentials.has(name) ? [] : [[name, value]];
-    }),
-  );
-  return {
-    ...inherited,
-    PYTHONPATH: [relayPath, inherited.PYTHONPATH].filter(Boolean).join(delimiter),
-    ...(egressAllowedHost
-      ? {
-          MAKA_EVAL_EGRESS_REQUIRED: '1',
-          MAKA_EVAL_EGRESS_ALLOWED_HOST: egressAllowedHost,
-        }
-      : {}),
-    ...(networkPolicyPath ? { MAKA_EVAL_NETWORK_POLICY_PATH: networkPolicyPath } : {}),
-  };
 }
 
 function mergeExecutionEnvironment(
